+++ tests/heart_rate_service_test.py
@@ -20,8 +20,7 @@ import pytest
 
 from bumble import device as device_module
 from bumble.profiles import heart_rate_service
-
-from . import test_utils
+from bumble.testing import test_utils
 
 
 # -----------------------------------------------------------------------------
